#47 - [BE][FEAT] paging for ratings get request

+++ backend/src/main/java/com/dacn/backend/dto/RatingRequestDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class RatingRequestDTO {
     private String userId;
     private String stationId;
-    private Double point;
+    private int point;
     private String comment;
 }
